+++ internal/api/handler/handler_err.go
@@ -1,0 +1,11 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/jbdoumenjou/go-rssaggregator/internal/api/json"
+)
+
+func ErrorHandler(w http.ResponseWriter, _ *http.Request) {
+	json.RespondWithError(w, http.StatusInternalServerError, http.StatusText(http.StatusInternalServerError))
+}
