+++ internal/api/handler/handler_readiness.go
@@ -1,0 +1,11 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/jbdoumenjou/go-rssaggregator/internal/api/json"
+)
+
+func ReadinessHandler(w http.ResponseWriter, _ *http.Request) {
+	json.RespondWithJSON(w, http.StatusOK, map[string]string{"status": "ok"})
+}
