+++ internal/api/handler/handler_user.go
@@ -1,4 +1,4 @@
-package api
+package handler
 
 import (
 	"database/sql"
@@ -6,6 +6,8 @@ import (
 	"errors"
 	"log/slog"
 	"net/http"
+
+	json2 "github.com/jbdoumenjou/go-rssaggregator/internal/api/json"
 
 	"github.com/google/uuid"
 
@@ -32,7 +34,7 @@ func (h *UserHandler) CreateUser(w http.ResponseWriter, r *http.Request) {
 	var req createUserReq
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
 		slog.Log(r.Context(), slog.LevelInfo, "decode user: %v", err)
-		respondWithError(w, http.StatusBadRequest, err.Error())
+		json2.RespondWithError(w, http.StatusBadRequest, err.Error())
 		return
 	}
 
@@ -40,10 +42,10 @@ func (h *UserHandler) CreateUser(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		// Error should be filtered here.
 		slog.Log(r.Context(), slog.LevelError, "create user: %v", err)
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		json2.RespondWithError(w, http.StatusInternalServerError, err.Error())
 	}
 
-	respondWithJSON(w, http.StatusOK, user)
+	json2.RespondWithJSON(w, http.StatusOK, user)
 }
 
 // GetUser get a user from the authorization header.
@@ -51,21 +53,21 @@ func (h *UserHandler) GetUser(w http.ResponseWriter, r *http.Request) {
 	userIDVal := r.Context().Value("user")
 	userID, ok := userIDVal.(uuid.UUID)
 	if userIDVal == nil || !ok {
-		respondWithError(w, http.StatusForbidden, http.StatusText(http.StatusForbidden))
+		json2.RespondWithError(w, http.StatusForbidden, http.StatusText(http.StatusForbidden))
 		return
 	}
 
 	user, err := h.db.GetUserFromId(r.Context(), userID)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
-			respondWithError(w, http.StatusForbidden, http.StatusText(http.StatusForbidden))
+			json2.RespondWithError(w, http.StatusForbidden, http.StatusText(http.StatusForbidden))
 			return
 		}
 		// Error should be filtered here.
 		slog.Log(r.Context(), slog.LevelError, "get user: %v", err)
-		respondWithError(w, http.StatusInternalServerError, err.Error())
+		json2.RespondWithError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
 
-	respondWithJSON(w, http.StatusOK, user)
+	json2.RespondWithJSON(w, http.StatusOK, user)
 }
