--- internal/api/handler_err.go
@@ -1,7 +1,0 @@
-package api
-
-import "net/http"
-
-func errorHandler(w http.ResponseWriter, _ *http.Request) {
-	respondWithError(w, http.StatusInternalServerError, http.StatusText(http.StatusInternalServerError))
-}
