--- internal/api/handler_readiness.go
@@ -1,7 +1,0 @@
-package api
-
-import "net/http"
-
-func readinessHandler(w http.ResponseWriter, _ *http.Request) {
-	respondWithJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-}
