+++ internal/api/router.go
@@ -3,6 +3,10 @@ package api
 import (
 	"net/http"
 
+	"github.com/jbdoumenjou/go-rssaggregator/internal/api/middleware"
+
+	"github.com/jbdoumenjou/go-rssaggregator/internal/api/handler"
+
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/cors"
 	"github.com/jbdoumenjou/go-rssaggregator/internal/database"
@@ -33,11 +37,11 @@ func NewRouter(db database.Querier) http.Handler {
 }
 
 func addV1Routes(r chi.Router, db database.Querier) {
-	r.Get("/readiness", readinessHandler)
-	r.Get("/err", errorHandler)
+	r.Get("/readiness", handler.ReadinessHandler)
+	r.Get("/err", handler.ErrorHandler)
 
-	userHandler := NewUserHandler(db)
-	authHandler := NewAuthMiddleware(db)
+	userHandler := handler.NewUserHandler(db)
+	middleware := middleware.NewAuthMiddleware(db)
 	r.Post("/users", userHandler.CreateUser)
-	r.Get("/users", authHandler.AuthMiddleware(userHandler.GetUser))
+	r.Get("/users", middleware.Authenticate(userHandler.GetUser))
 }
